refactoring JS
